Move strut layout to line_break module
