+++ weasyprint/css/computed_values.py
@@ -5,11 +5,10 @@ from math import pi
 from tinycss2.color4 import parse_color
 
 from ..logger import LOGGER
-from ..text.ffi import FROM_UNITS, ffi, pango
-from ..text.line_break import Layout, first_line_metrics
+from ..text.line_break import strut
 from ..urls import get_link_attribute, get_url_tuple
 from .properties import INITIAL_VALUES, ZERO_PIXELS, Dimension
-from .units import ANGLE_TO_RADIANS, LENGTH_UNITS, LENGTHS_TO_PIXELS
+from .units import ANGLE_TO_RADIANS, LENGTH_UNITS, to_pixels
 
 # Value in pixels of font-size for <absolute-size> keywords: 12pt (16px) for
 # medium, and scaling factors given in CSS3 for others:
@@ -120,34 +119,11 @@ PAGE_SIZES = {
 assert all(width.value < height.value for width, height in PAGE_SIZES.values())
 
 INITIAL_PAGE_SIZE = PAGE_SIZES['a4']
-INITIAL_VALUES['size'] = tuple(
-    size.value * LENGTHS_TO_PIXELS[size.unit] for size in INITIAL_PAGE_SIZE)
+INITIAL_VALUES['size'] = tuple(to_pixels(size) for size in INITIAL_PAGE_SIZE)
 
 
 # Maps property names to functions returning the computed values
 COMPUTER_FUNCTIONS = {}
-
-
-def _font_style_cache_key(style, include_size=False):
-    key = str((
-        style['font_family'],
-        style['font_style'],
-        style['font_stretch'],
-        style['font_weight'],
-        style['font_variant_ligatures'],
-        style['font_variant_position'],
-        style['font_variant_caps'],
-        style['font_variant_numeric'],
-        style['font_variant_alternates'],
-        style['font_variant_east_asian'],
-        style['font_feature_settings'],
-        style['font_variation_settings'],
-        style['font_language_override'],
-        style['lang'],
-    ))
-    if include_size:
-        key += str(style['font_size']) + str(style['line_height'])
-    return key
 
 
 def register_computer(name):
@@ -268,40 +244,14 @@ def length(style, name, value, font_size=None, pixels_only=False):
     """Compute a length ``value``."""
     if value in ('auto', 'content', 'from-font'):
         return value
-    if value.value == 0:
+    elif value.value == 0:
         return 0 if pixels_only else ZERO_PIXELS
-
-    unit = value.unit
-    if unit == 'px':
-        return value.value if pixels_only else value
-    elif unit in LENGTHS_TO_PIXELS:
-        # Convert absolute lengths to pixels
-        result = value.value * LENGTHS_TO_PIXELS[unit]
-    elif unit in ('em', 'ex', 'ch', 'rem', 'lh', 'rlh'):
-        if font_size is None:
-            font_size = style['font_size']
-        if unit == 'ex':
-            # TODO: use context to use @font-face fonts
-            ratio = character_ratio(style, 'x')
-            result = value.value * font_size * ratio
-        elif unit == 'ch':
-            ratio = character_ratio(style, '0')
-            result = value.value * font_size * ratio
-        elif unit == 'em':
-            result = value.value * font_size
-        elif unit == 'rem':
-            result = value.value * style.root_style['font_size']
-        elif unit == 'lh':
-            line_height, _ = strut_layout(style)
-            result = value.value * line_height
-        elif unit == 'rlh':
-            line_height, _ = strut_layout(style.root_style)
-            result = value.value * line_height
-    else:
+    elif value.unit not in LENGTH_UNITS:
         # A percentage or 'auto': no conversion needed.
         return value
 
-    return result if pixels_only else Dimension(result, 'px')
+    pixels = to_pixels(value, style, font_size)
+    return pixels if pixels_only else Dimension(pixels, 'px')
 
 
 @register_computer('bleed-left')
@@ -745,7 +695,7 @@ def vertical_align(style, name, value):
     elif value == 'sub':
         return style['font_size'] * -0.5
     elif value.unit == '%':
-        height, _ = strut_layout(style)
+        height, _ = strut(style)
         return height * value.value / 100
     else:
         return length(style, name, value, pixels_only=True)
@@ -758,75 +708,3 @@ def word_spacing(style, name, value):
         return 0
     else:
         return length(style, name, value, pixels_only=True)
-
-
-def strut_layout(style, context=None):
-    """Return a tuple of the used value of ``line-height`` and the baseline.
-
-    The baseline is given from the top edge of line height.
-
-    """
-    if style['font_size'] == 0:
-        return 0, 0
-
-    if context:
-        key = _font_style_cache_key(style, include_size=True)
-        if key in context.strut_layouts:
-            return context.strut_layouts[key]
-
-    layout = Layout(context, style)
-    layout.set_text(' ')
-    line, _ = layout.get_first_line()
-    _, _, _, _, text_height, baseline = first_line_metrics(
-        line, '', layout, resume_at=None, space_collapse=False, style=style)
-    if style['line_height'] == 'normal':
-        result = text_height, baseline
-        if context:
-            context.strut_layouts[key] = result
-        return result
-    type_, line_height = style['line_height']
-    if type_ == 'NUMBER':
-        line_height *= style['font_size']
-    result = line_height, baseline + (line_height - text_height) / 2
-    if context:
-        context.strut_layouts[key] = result
-    return result
-
-
-def character_ratio(style, character):
-    """Return the ratio of 1ex/font_size or 1ch/font_size."""
-    # TODO: use context to use @font-face fonts
-
-    assert character in ('x', '0')
-
-    cache = style.cache[f'ratio_{"ex" if character == "x" else "ch"}']
-    cache_key = _font_style_cache_key(style)
-    if cache_key in cache:
-        return cache[cache_key]
-
-    # Avoid recursion for letter-spacing and word-spacing properties
-    style = style.copy()
-    style['letter_spacing'] = 'normal'
-    style['word_spacing'] = 0
-    # Random big value
-    style['font_size'] = 1000
-
-    layout = Layout(context=None, style=style)
-    layout.set_text(character)
-    line, _ = layout.get_first_line()
-
-    ink_extents = ffi.new('PangoRectangle *')
-    logical_extents = ffi.new('PangoRectangle *')
-    pango.pango_layout_line_get_extents(line, ink_extents, logical_extents)
-    if character == 'x':
-        measure = -ink_extents.y * FROM_UNITS
-    else:
-        measure = logical_extents.width * FROM_UNITS
-    ffi.release(ink_extents)
-    ffi.release(logical_extents)
-
-    # Zero means some kind of failure, fallback is 0.5.
-    # We round to try keeping exact values that were altered by Pango.
-    ratio = round(measure / style['font_size'], 5) or 0.5
-    cache[cache_key] = ratio
-    return ratio
