+++ weasyprint/layout/inline.py
@@ -4,9 +4,7 @@ import unicodedata
 from math import inf
 
 from ..css import computed_from_cascaded
-from ..css.computed_values import character_ratio, strut_layout
 from ..formatting_structure import boxes, build
-from ..text.line_break import can_break_text, create_layout, split_first_line
 from .absolute import AbsolutePlaceholder, absolute_layout
 from .flex import flex_layout
 from .float import avoid_collisions, float_layout
@@ -17,6 +15,9 @@ from .percent import resolve_one_percentage, resolve_percentages
 from .preferred import inline_min_content_width, shrink_to_fit, trailing_whitespace_size
 from .replaced import inline_replaced_box_layout
 from .table import find_in_flow_baseline, table_wrapper_width
+
+from ..text.line_break import (  # isort:skip
+    can_break_text, character_ratio, create_layout, split_first_line, strut)
 
 
 def iter_line_boxes(context, box, position_y, bottom_space, skip_stack,
@@ -75,7 +76,7 @@ def get_next_linebox(context, linebox, position_y, bottom_space, skip_stack,
         # Width and height must be calculated to avoid floats
         linebox.width = inline_min_content_width(
             context, linebox, skip_stack=skip_stack, first_line=True)
-        linebox.height, _ = strut_layout(linebox.style, context)
+        linebox.height, _ = strut(linebox.style, context)
     else:
         # No float, width and height will be set by the lines
         linebox.width = linebox.height = 0
@@ -858,7 +859,7 @@ def split_inline_box(context, box, position_x, max_x, bottom_space, skip_stack,
         new_box.width = position_x - content_box_left
         new_box.translate(dx=float_widths['left'], ignore_floats=True)
 
-    line_height, new_box.baseline = strut_layout(box.style, context)
+    line_height, new_box.baseline = strut(box.style, context)
     new_box.height = box.style['font_size']
     half_leading = (line_height - new_box.height) / 2
     # Set margins to the half leading but also compensate for borders and
@@ -922,7 +923,7 @@ def split_text_box(context, box, available_width, skip, is_line_start=True):
         # "only the 'line-height' is used when calculating the height
         #  of the line box."
         # Set margins so that margin_height() == line_height
-        line_height, _ = strut_layout(box.style, context)
+        line_height, _ = strut(box.style, context)
         half_leading = (line_height - height) / 2
         box.margin_top = half_leading
         box.margin_bottom = half_leading
